backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Limiter==3.7.0
Flask-Mail==0.10.0
Werkzeug==3.0.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
alembic==1.13.1
bcrypt==4.1.3
python-jose==3.3.0
bleach==6.1.0
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
cloudinary==1.40.0
Pillow==10.3.0
redis==5.0.4
elasticsearch==8.13.0
requests==2.31.0
gunicorn==22.0.0
python-dotenv==1.0.1
phonenumbers==8.13.36
pyyaml==6.0.2
